+++ src/main/java/com/management/EmpManagement/Dto/request/HalfDayReqDto.java
@@ -1,4 +1,4 @@
-package com.management.EmpManagement.DTO.response;
+package com.management.EmpManagement.Dto.request;
 
 import com.management.EmpManagement.Enum.Status;
 import com.management.EmpManagement.Entity.Employee;
@@ -8,19 +8,16 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.sql.Date;
-import java.sql.Time;
-
 
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
 @Builder
 
-public class PartialResDto {
-    private Long pwId;
+public class HalfDayReqDto {
+
     private Status leaveType;
     private String reason;
     private Date ApplyDate;
-    private Time timeOfPartialDay;
     private Employee employee;
 }
