+++ src/main/java/com/management/EmpManagement/Dto/request/PartialReqDto.java
@@ -1,4 +1,4 @@
-package com.management.EmpManagement.DTO.request;
+package com.management.EmpManagement.Dto.request;
 
 import com.management.EmpManagement.Enum.Status;
 import com.management.EmpManagement.Entity.Employee;
@@ -8,16 +8,18 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.sql.Date;
+import java.sql.Time;
 
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
 @Builder
 
-public class HalfDayReqDto {
+public class PartialReqDto {
 
     private Status leaveType;
     private String reason;
     private Date ApplyDate;
+    private Time timeOfPartialDay;
     private Employee employee;
 }
