+++ src/main/java/com/management/EmpManagement/Dto/response/HalfDayResDto.java
@@ -1,4 +1,4 @@
-package com.management.EmpManagement.DTO.request;
+package com.management.EmpManagement.Dto.response;
 
 import com.management.EmpManagement.Enum.Status;
 import com.management.EmpManagement.Entity.Employee;
@@ -9,18 +9,16 @@ import lombok.NoArgsConstructor;
 
 import java.sql.Date;
 
-
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
 @Builder
 
-public class WorkFromHomeReqDto {
-
+public class HalfDayResDto {
+    private Long pwId;
     private Status leaveType;
     private String reason;
     private Date ApplyDate;
-    private Date startDate;
-    private Date endDate;
     private Employee employee;
+
 }
